Move RemoteFacilitiesViewset to core to be used by other parts of the application
