+++ kolibri/core/device/api_urls.py
@@ -10,6 +10,7 @@ from .api import DeviceRestartView
 from .api import DeviceSettingsView
 from .api import DriveInfoViewSet
 from .api import FreeSpaceView
+from .api import RemoteFacilitiesViewset
 from .api import UserSyncStatusViewSet
 
 router = routers.SimpleRouter()
@@ -32,4 +33,7 @@ urlpatterns = [
     url(r"^devicesettings/", DeviceSettingsView.as_view(), name="devicesettings"),
     url(r"^devicename/", DeviceNameView.as_view(), name="devicename"),
     url(r"^devicerestart/", DeviceRestartView.as_view(), name="devicerestart"),
+    url(
+        r"^remotefacilities", RemoteFacilitiesViewset.as_view(), name="remotefacilities"
+    ),
 ]
